Rewrote LPD8806 Ethernet sketch based on etherCard
This rewrite removes the dependency on the Adafruit LPD8806 library, because while it is useful, it's also a waste of resources for this particular application.

The etherCard libary makes the sketch itself significantly easier to read, removing the old packetloop to a single library funcion that seems comparatively massively optimized with regard to memory.

The ethernet buffer has been reduced to the very minimum necessary for the application, clocking in at 534 bytes.