Placed try/except on the correct level.
The try: statement is now on the inside of the loop. This means that if a single position is missed but there are more to come, these will also be accessed. That is, if only LED 4 (out of 160) is not present, the old could would place 'AAA' on the fourth position and then stop. The current code places 'AAA' on that position and carries on processing the following 155 LED's.